requirements.txt
zirconium>=1.2.2
autoinject>=1.3.2
zrlog>=0.3.0
graceful_shutdown
click
pyyaml
toml
flask
prometheus_client
requests
aiohttp
azure-identity
azure-storage-blob
